This is a fix attempt and spec for an issue seen in the iOS example: distanceTraveled.
Essentially CLLocationManager has a property named locationServicesEnabled, and a query method with the same name. If we try to call the query method, previously we'd comvert that to property access and break compilation. This fix attempt basically skips attemtping to determine static field access if the parent of the Dot node is a Call.

This is just one fix attempt, it may actually break other uses cases where we do want to access static properties as part of a method call chain - I'm not entirely sure what we might want to test to ensure it doesn't break other scenarios.